+++ src/openrct2/interface/Colour.cpp
@@ -57,6 +57,49 @@ void colours_init_maps()
     }
 }
 
+namespace Colour
+{
+    colour_t FromString(const std::string_view& s, colour_t defaultValue)
+    {
+        static const std::unordered_map<std::string_view, colour_t> LookupTable{
+            { "black", COLOUR_BLACK },
+            { "grey", COLOUR_GREY },
+            { "white", COLOUR_WHITE },
+            { "dark_purple", COLOUR_DARK_PURPLE },
+            { "light_purple", COLOUR_LIGHT_PURPLE },
+            { "bright_purple", COLOUR_BRIGHT_PURPLE },
+            { "dark_blue", COLOUR_DARK_BLUE },
+            { "light_blue", COLOUR_LIGHT_BLUE },
+            { "icy_blue", COLOUR_ICY_BLUE },
+            { "teal", COLOUR_TEAL },
+            { "aquamarine", COLOUR_AQUAMARINE },
+            { "saturated_green", COLOUR_SATURATED_GREEN },
+            { "dark_green", COLOUR_DARK_GREEN },
+            { "moss_green", COLOUR_MOSS_GREEN },
+            { "bright_green", COLOUR_BRIGHT_GREEN },
+            { "olive_green", COLOUR_OLIVE_GREEN },
+            { "dark_olive_green", COLOUR_DARK_OLIVE_GREEN },
+            { "bright_yellow", COLOUR_BRIGHT_YELLOW },
+            { "yellow", COLOUR_YELLOW },
+            { "dark_yellow", COLOUR_DARK_YELLOW },
+            { "light_orange", COLOUR_LIGHT_ORANGE },
+            { "dark_orange", COLOUR_DARK_ORANGE },
+            { "light_brown", COLOUR_LIGHT_BROWN },
+            { "saturated_brown", COLOUR_SATURATED_BROWN },
+            { "dark_brown", COLOUR_DARK_BROWN },
+            { "salmon_pink", COLOUR_SALMON_PINK },
+            { "bordeaux_red", COLOUR_BORDEAUX_RED },
+            { "saturated_red", COLOUR_SATURATED_RED },
+            { "bright_red", COLOUR_BRIGHT_RED },
+            { "dark_pink", COLOUR_DARK_PINK },
+            { "bright_pink", COLOUR_BRIGHT_PINK },
+            { "light_pink", COLOUR_LIGHT_PINK },
+        };
+        auto result = LookupTable.find(s);
+        return (result != LookupTable.end()) ? result->second : defaultValue;
+    }
+} // namespace Colour
+
 #ifndef NO_TTF
 static uint8_t BlendColourMap[PALETTE_COUNT][PALETTE_COUNT] = { 0 };
