+++ src/openrct2/object/Object.cpp
@@ -37,6 +37,21 @@ void Object::ReadLegacy(IReadObjectContext* context, OpenRCT2::IStream* stream)
     throw std::runtime_error("Not supported.");
 }
 
+void Object::PopulateTablesFromJson(IReadObjectContext* context, json_t& root)
+{
+    _stringTable.ReadJson(root);
+    _imageTable.ReadJson(context, root);
+}
+
+rct_object_entry Object::ParseObjectEntry(const std::string& s)
+{
+    rct_object_entry entry = {};
+    std::fill_n(entry.name, sizeof(entry.name), ' ');
+    auto copyLen = std::min<size_t>(8, s.size());
+    std::copy_n(s.c_str(), copyLen, entry.name);
+    return entry;
+}
+
 std::string Object::GetOverrideString(uint8_t index) const
 {
     auto legacyIdentifier = GetLegacyIdentifier();
